Revamped the entire benchmarks directory structure.

+++ include/benchmarks/gemm_blaze.h
@@ -7,12 +7,12 @@
 #include "random.hpp"
 
 static void gemm_blaze(benchmark::State& state) {
-    blaze::StaticMatrix<float, SIZE, SIZE, blaze::rowMajor> m1, m2, res;
+    blaze::StaticMatrix<double, SIZE, SIZE, blaze::rowMajor> m1, m2, res;
 
     m1 =
-        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomFloat(0, 100'000.0); });
+        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomDouble(0, 100'000.0); });
     m2 =
-        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomFloat(0, 100'000.0); });
+        blaze::generate(SIZE, SIZE, [](std::size_t i, std::size_t j) { return randomDouble(0, 100'000.0); });
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(res);
