+++ include/benchmarks/inversion_blaze.h
@@ -7,7 +7,7 @@
 #include "random.hpp"
 
 static void inversion_blaze(benchmark::State& state) {
-    blaze::StaticMatrix<float, SIZE, SIZE, blaze::rowMajor> m;
+    blaze::StaticMatrix<double, SIZE, SIZE, blaze::rowMajor> m;
 
     for (int i = 0; i < SIZE; i++) {
         for (int j = 0; j < SIZE - i; j++) {
@@ -16,7 +16,7 @@ static void inversion_blaze(benchmark::State& state) {
         }
     }
 
-	blaze::SymmetricMatrix<blaze::StaticMatrix<float, SIZE, SIZE, blaze::rowMajor>> m1(m), res;
+	blaze::SymmetricMatrix<blaze::StaticMatrix<double, SIZE, SIZE, blaze::rowMajor>> m1(m), res;
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(res);
